Add support for & examples of translatable strings with context...
... This includes functions callable from Nyquist but not yet any examples
of strings with context in any .ny files.

This does include some examples of strings with context, and with both plurals
and context, in C++ files.  It is not intended to fix possible ambiguities
exhaustively.

Note that wxWidgets 3.1.1 is sufficient to support contexts.

If Audacity is build with wxWidgets 3.0, it will run, but strings with
context will fail to translate.